Change retention of ignition data from 7 to 60 days (#18)
This is done to support service data reminders. In order to make sure
that service data is not marked as too old for users on vacation or cases
where the car is not used and hence do not send data, we need to check
the ignition to see if the car has been driven for this period.